h3_quarter_1/i3.4_Client-Server Applications/Урок-4/lesson_4_Tynnikov/lesson_4_Tynnikov/client.py
"""Программа-клиент"""

import sys
import json
import socket
import time
from common.variables import ACTION, PRESENCE, TIME, USER, ACCOUNT_NAME, \
    RESPONSE, ERROR, DEFAULT_IP_ADDRESS, DEFAULT_PORT
from common.utils import get_message, send_message


def create_presence(account_name='Guest'):
    '''
    Функция генерирует запрос о присутствии клиента
    :param account_name:
    :return:
    '''
    # {'action': 'presence', 'time': 1573760672.167031, 'user': {'account_name': 'Guest'}}
    out = {
        ACTION: PRESENCE,
        TIME: time.time(),
        USER: {
            ACCOUNT_NAME: account_name
        }
    }
    return out


def process_ans(message):
    '''
    Функция разбирает ответ сервера
    :param message:
    :return:
    '''
    if RESPONSE in message:
        if message[RESPONSE] == 200:
            return '200 : OK'
        return f'400 : {ERROR}'
    raise ValueError


def get_client_address_and_port():
    try:
        server_address = sys.argv[1]
        server_port = int(sys.argv[2])
        if server_port < 1024 or server_port > 65535:
            raise ValueError
        return server_address, server_port
    except IndexError:
        server_address = DEFAULT_IP_ADDRESS
        server_port = DEFAULT_PORT
        return server_address, server_port
    except ValueError:
        print('В качестве порта может быть указано только число в диапазоне от 1024 до 65535.')
        sys.exit(1)


def main():
    """
    Загружаем параметы коммандной строки.
    Настройка пайчарм конфига для client.py 192.168.0.113 8079
    :return: None
    """

    server_address, server_port = get_client_address_and_port()

    client_socket = socket.socket(socket.AF_INET, socket.SOCK_STREAM)  # Инициализация сокета и обмен
    client_socket.connect((server_address, server_port))
    message_to_server = create_presence()
    send_message(client_socket, message_to_server)
    try:
        answer = process_ans(get_message(client_socket))
        print(answer)
    except (ValueError, json.JSONDecodeError):
        print('Не удалось декодировать сообщение сервера.')


if __name__ == '__main__':
    main()
